feat(appointments): correct issues revealed by integration tests    - Enforce user authorization on update and delete operations    - Set correct HTTP status 201 when creating a new appointment    - Remove deprecated handleFindByAppointmentDate method    - Update use-case and repository types to reflect new data structure

+++ source/back-end/src/services/appointments.use-case.ts
@@ -1,5 +1,5 @@
 import { type Prisma, type Appointment } from '@prisma/client'
-import { type AppointmentRepository } from '../repository/protocols/appointment.repository'
+import { type FindByIdAppointments, type AppointmentRepository } from '../repository/protocols/appointment.repository'
 import { RecordExistence } from '../utils/validation/record-existence.validation.util'
 import { type CustomerRepository } from '../repository/protocols/customer.repository'
 import { type ProfessionalRepository } from '../repository/protocols/professional.repository'
@@ -22,12 +22,11 @@ class AppointmentsUseCase {
 
   public async executeFindAll (): Promise<AppointmentOutput> {
     const appointments = await this.appointmentRepository.findAll()
-    RecordExistence.validateManyRecordsExistence(appointments, 'appointments')
 
     return { appointments }
   }
 
-  public async executeFindById (appointmentId: string): Promise<Appointment | null> {
+  public async executeFindById (appointmentId: string): Promise<FindByIdAppointments | null> {
     const appointment = await this.appointmentRepository.findById(appointmentId)
     RecordExistence.validateRecordExistence(appointment, this.entityName)
 
@@ -46,16 +45,8 @@ class AppointmentsUseCase {
     return { appointments }
   }
 
-  public async executeFindByAppointmentDate (appointmentDate: Date): Promise<AppointmentOutput> {
-    const appointments = await this.appointmentRepository.findByAppointmentDate(appointmentDate)
-    RecordExistence.validateManyRecordsExistence(appointments, 'appointments')
-
-    return { appointments }
-  }
-
   public async executeFindByServiceOfferedId (appointmentId: string): Promise<AppointmentOutput> {
     const appointments = await this.appointmentRepository.findByServiceOfferedId(appointmentId)
-    RecordExistence.validateManyRecordsExistence(appointments, 'appointments')
 
     return { appointments }
   }
@@ -87,15 +78,24 @@ class AppointmentsUseCase {
     return newAppointment
   }
 
-  public async executeUpdate (appointmentId: string, appointmentToUpdate: Prisma.AppointmentUpdateInput) {
-    await this.executeFindById(appointmentId)
+  public async executeUpdate (userId: string, appointmentId: string, appointmentToUpdate: Prisma.AppointmentUpdateInput) {
+    const appointmentById = await this.executeFindById(appointmentId)
+
+    if (appointmentById?.customerId !== userId && appointmentById?.offer.professionalId !== userId) {
+      throw new CustomError('You are not allowed to update this appointment', 403)
+    }
+
     const updatedCustomer = await this.appointmentRepository.update(appointmentId, appointmentToUpdate)
 
     return updatedCustomer
   }
 
-  public async executeDelete (appointmentId: string) {
-    await this.executeFindById(appointmentId)
+  public async executeDelete (userId: string, appointmentId: string) {
+    const appointment = await this.executeFindById(appointmentId)
+    if (appointment?.customerId !== userId && appointment?.offer.professionalId !== userId) {
+      throw new CustomError('You are not allowed to delete this appointment', 403)
+    }
+
     const deletedAppointment = await this.appointmentRepository.delete(appointmentId)
 
     return deletedAppointment
